+++ src/commands/addItem/addStructCommand.ts
@@ -1,10 +1,8 @@
 import * as vscode from 'vscode';
-import { injectable } from 'inversify';
-import { TemplateType } from '../templates/templateType';
-import { AddTemplateCommand } from './addTemplateCommand';
+import { TemplateType } from '../../templates/templateType';
+import { AddItemCommand } from './addItemCommand';
 
-@injectable()
-export class AddStructCommand extends AddTemplateCommand {
+export class AddStructCommand extends AddItemCommand {
 
     public readonly id: string = 'c-sharp-utilities.addStruct';