+++ src/commands/index.ts
@@ -1,13 +1,19 @@
-import { AddClassCommand } from "./addClassCommand";
-import { AddEnumCommand } from "./addEnumCommand";
-import { AddInterfaceCommand } from "./addInterfaceCommand";
-import { AddRecordCommand } from "./addRecordCommand";
-import { AddStructCommand } from "./addStructCommand";
+import { AddClassCommand } from "./addItem/addClassCommand";
+import { AddEnumCommand } from "./addItem/addEnumCommand";
+import { AddInterfaceCommand } from "./addItem/addInterfaceCommand";
+import { AddRecordCommand } from "./addItem/addRecordCommand";
+import { AddStructCommand } from "./addItem/addStructCommand";
+import { ManageProjectReferencesCommand } from "./manageProjectReferencesCommand";
+import { RefreshProjectReferenceTreeViewCommand } from "./projectReference/refreshProjectReferenceTreeViewCommand";
+import { SelectRootProjectCommand } from "./selectRootProjectCommand";
 
 export {
     AddClassCommand,
     AddEnumCommand,
     AddInterfaceCommand,
     AddRecordCommand,
-    AddStructCommand
+    AddStructCommand,
+    ManageProjectReferencesCommand,
+    SelectRootProjectCommand,
+    RefreshProjectReferenceTreeViewCommand
 };