+++ src/handlers/cSharpProject.ts
@@ -1,5 +1,10 @@
+import * as vscode from 'vscode';
+
 export interface CSharpProject {
+    name: string;
+    uri: vscode.Uri;
     path: string;
     rootNamespace: string;
     projectReferencePaths: string[];
+    projectReferenceUris: vscode.Uri[];
 }